Fix endpoint layout

+++ src/main/java/com/company/enroller/controllers/MeetingRestController.java
@@ -30,44 +30,44 @@ public class MeetingRestController {
     @RequestMapping(value = "", method = RequestMethod.GET)
     public ResponseEntity<?> getMeetings() {
         Collection<Meeting> meetings = meetingService.getAll();
-        return new ResponseEntity<Collection<Meeting>>(meetings, HttpStatus.OK);
+        return new ResponseEntity<>(meetings, HttpStatus.OK);
     }
 
     // TODO refactor to use exceptions to simplify this method conditional statements
     @RequestMapping(value = "", method = RequestMethod.POST)
     public ResponseEntity<?> addMeeting(@RequestBody Meeting meeting) {
     	if (meetingService.get(meeting.getId()) != null)
-    		return new ResponseEntity<String>("Meeting already exist", HttpStatus.CONFLICT);
+    		return new ResponseEntity<>("Meeting already exist", HttpStatus.CONFLICT);
     	if (meetingService.add(meeting) == null) 
-    		return new ResponseEntity<String>("Unable to add meeting", HttpStatus.INTERNAL_SERVER_ERROR);
-    	return new ResponseEntity<Meeting>(meeting, HttpStatus.CREATED);
+    		return new ResponseEntity<>("Unable to add meeting", HttpStatus.INTERNAL_SERVER_ERROR);
+    	return new ResponseEntity<>(meeting, HttpStatus.CREATED);
     }
 
     // TODO refactor to use exceptions to simplify this method conditional statements
-    @RequestMapping(value = "/{id}", method = RequestMethod.POST) 
+    @RequestMapping(value = "/{id}/participants", method = RequestMethod.POST)
     public ResponseEntity<?> addParticipant(@PathVariable("id") long meetingId, @RequestBody String participantLogin) {
     	Participant participant = participantService.findByLogin(participantLogin);
     	Meeting meeting = meetingService.get(meetingId);
     	if (meeting == null) {
-    		return new ResponseEntity<String>("Meeting doesn't exist", HttpStatus.BAD_REQUEST);
+    		return new ResponseEntity<>("Meeting doesn't exist", HttpStatus.BAD_REQUEST);
     	}
     	if (!meeting.hasParticipant(participant)) {
             Meeting result = (Meeting) meetingService.addParticipant(meeting, participant);
             if (result == null) {
-                return new ResponseEntity<String>("Unable to add participant", HttpStatus.INTERNAL_SERVER_ERROR);
+                return new ResponseEntity<>("Unable to add participant", HttpStatus.INTERNAL_SERVER_ERROR);
             }
-            return new ResponseEntity<Meeting>(result, HttpStatus.CREATED);
+            return new ResponseEntity<>(result, HttpStatus.CREATED);
         } else {
     	    // particpant already enrolled
-            return new ResponseEntity<String>("Participant already enrolled", HttpStatus.CONFLICT);
+            return new ResponseEntity<>("Participant already enrolled", HttpStatus.CONFLICT);
         }
     }
 
-    @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
-    public ResponseEntity<?> removeParticipant(@PathVariable("id") long meetingId, @RequestBody String login) {
+    @RequestMapping(value = "/{id}/participants/{login}", method = RequestMethod.DELETE)
+    public ResponseEntity<?> removeParticipant(@PathVariable("id") long meetingId, @PathVariable String login) {
         try {
             Meeting updatedMeeting = meetingService.removeParticipant(meetingId, login);
-            return new ResponseEntity<Meeting>(updatedMeeting, HttpStatus.OK);
+            return new ResponseEntity<>(updatedMeeting, HttpStatus.OK);
         } catch (HibernateException e) {
             return new ResponseEntity<String>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -75,6 +75,8 @@ public class MeetingRestController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<?> deleteMeeting(@PathVariable("id") long meetingId) {
+        if(!meetingService.get(meetingId).getParticipants().isEmpty())
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         try {
             meetingService.delete(meetingId);
             return new ResponseEntity<>(HttpStatus.OK);
